+++ README.md
@@ -5,6 +5,7 @@
 - [x] 계산 결과에서 지우기 버튼을 눌렀을 때 동작 수정
 - [x] MVP 패턴으로 리팩터링
 - [x] Presenter 테스트 코드 구현
+- [x] 3차 피드백 반영
 
 ### 4단계 - UI 테스트
 - [x] 1차 피드백 반영
