--- app/src/main/java/camp/nextstep/edu/calculator/CalculatorViewInterface.kt
@@ -1,10 +1,0 @@
-package camp.nextstep.edu.calculator
-
-
-interface CalculatorViewInterface {
-    var presenter: CalculatorPresenterInterface
-
-    fun showCalculationResult(result: String)
-    fun alertIncompleteExpression()
-    fun showCalculationExpression(expression: String)
-}
